CV/GCF/requirements.txt
functions-framework>=3.0.0
GitPython>=3.1.20
Markdown>=3.3.0
google-cloud-secret-manager>=2.16.0
